changed function defenitions, moved mouse event fucntions to eperate header file
