+++ detection.cpp
@@ -1,15 +1,9 @@
 #include <opencv2/opencv.hpp>
-#include<X11/Xlib.h>
-#include <X11/extensions/XTest.h>
+#include"Mouse.h"
 
 using namespace cv;
 
-void setCursor(int x, int y){
-  Display *dpy = XOpenDisplay(NULL); // open the default display
-  Window root = DefaultRootWindow(dpy); // get the root window
-  XWarpPointer(dpy, None, root, 0, 0, 0, 0, x, y); // move the pointer
-  XCloseDisplay(dpy);
-}
+
 
 double findPointsDistance(Point a, Point b) {
     Point difference = a - b;
@@ -23,17 +17,10 @@ double findAngle(Point a, Point b, Point c) {
     return acos((bc * bc + ca * ca - ab * ab) / (2 * bc * ca));
 }
 
-void click(int button, bool state){
-    Display* dpy = XOpenDisplay(NULL);
-    // button = left mouse button and button = 2 for middle button and button = 3 for right button.
-    // state = True implies button press, state = false implies button release
-    XTestFakeButtonEvent(dpy, button, state, 0); // Press or Release Mouse Button
-    XFlush(dpy);
-    XCloseDisplay(dpy);
-}
 
 
-Point detectHands(Mat *frame,Mat *background,Mat *binn){
+
+void detectHands(Mat *frame,Mat *background,Mat *binn){
     Mat one,two;
     cvtColor(*frame, one, COLOR_BGR2GRAY);
     cvtColor(*background, two, COLOR_BGR2GRAY);
@@ -124,5 +111,5 @@ Point detectHands(Mat *frame,Mat *background,Mat *binn){
         setCursor(topX*3, topY*3);
     }
 
-    return topPoint;   
+
 }
